requirements.txt
annotated-types          0.7.0
blinker                  1.8.2
cachetools               5.3.3
certifi                  2024.2.2
charset-normalizer       3.3.2
click                    8.1.7
colorama                 0.4.6
dnspython                2.6.1
Flask                    3.0.3
google-api-core          2.19.0
google-api-python-client 2.129.0
google-auth              2.29.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.0
googleapis-common-protos 1.63.0
gspread                  6.1.2
httplib2                 0.22.0
idna                     3.7
itsdangerous             2.2.0
Jinja2                   3.1.4
MarkupSafe               2.1.5
numpy                    1.26.4
oauth2client             4.1.3
oauthlib                 3.2.2
pandas                   2.2.2
pip                      24.0
proto-plus               1.23.0
protobuf                 4.25.3
pyasn1                   0.6.0
pyasn1_modules           0.4.0
pydantic_core            2.18.2
pymongo                  4.7.2
pyparsing                3.1.2
python-dateutil          2.9.0.post0
python-decouple          3.8
python-dotenv            1.0.1
pytz                     2024.1
requests                 2.32.1
requests-oauthlib        2.0.0
rsa                      4.9
setuptools               65.5.1
six                      1.16.0
typing_extensions        4.11.0
tzdata                   2024.1
uritemplate              4.1.1
urllib3                  2.2.1
Werkzeug                 3.0.3
wheel                    0.38.4